feat: Message sending but no reception yet
feat: Unread message notification also sent to client in ConversationsAndUsersLists

feat: Store message draft in Convesation.Draft

refac: Rename EncryptedMessageContent to EncryptedMessageCopy and its receiver_id column to recipient_id

refac: Rename EcryptedAttachmentContent to EncryptedAttachmentCopy and its receiver_id column to recipient_id

feat: Remove Attachment.type database column

refac: Move AddParticipantViewModel to ViewModel\Conversations

fix: Prevent anyone from adding conversation owner as participant in AddParticipation

refac: Change unused PacketReader.ReadInt32 to useful ReadUInt32

+++ Shared/MVVM/Model/Networking/Transfer/Reception/PacketReader.cs
@@ -109,11 +109,11 @@ namespace Shared.MVVM.Model.Networking.Transfer.Reception
             return ret;
         }
 
-        public int ReadInt32()
+        public uint ReadUInt32()
         {
-            int ret = IPAddress.NetworkToHostOrder(
+            uint ret = (uint)IPAddress.NetworkToHostOrder(
                 BitConverter.ToInt32(_packet, _byteIndex));
-            Proceed(sizeof(int));
+            Proceed(sizeof(uint));
             return ret;
         }
